second phase
