+++ fe/spark-dpp/src/main/java/org/apache/doris/load/loadv2/dpp/DorisBucketPartitioner.java
@@ -1,0 +1,41 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.apache.doris.load.loadv2.dpp;
+
+import java.util.List;
+import java.util.Map;
+
+public class DorisBucketPartitioner extends DorisPartitioner {
+
+    private final Map<String, Integer> bucketKeyMap;
+
+    public DorisBucketPartitioner(Map<String, Integer> bucketKeyMap) {
+        this.bucketKeyMap = bucketKeyMap;
+    }
+
+    @Override
+    public int numPartitions() {
+        return bucketKeyMap.size();
+    }
+
+    @Override
+    public int getPartition(Object key) {
+        List<Object> rddKey = (List<Object>) key;
+        return bucketKeyMap.get(String.valueOf(rddKey.get(0)));
+    }
+}
